:sparkles: s7: Coffee stores lib
